Move Comment on Analysis dialog to presenter
